+++ CARET_trace/src/tracing_controller.cpp
@@ -29,36 +29,7 @@
 #include <unordered_set>
 #include <utility>
 
-#include <execinfo.h>
-#include <cstring>
-#include <unistd.h>
-static void extractc_fn(const char* symbol, char* fn, size_t bufSize) {
-    const char* start = strchr(symbol, '(');
-    if (start) {
-        ++start; // '('を飛ばす
-        const char* end = strchr(start, '+');
-        if (end) {
-            strncpy(fn, start, end - start);
-            fn[end - start] = '\0'; // 終端文字を追加
-        } else {
-            strncpy(fn, start, bufSize - 1);
-            fn[bufSize - 1] = '\0';
-        }
-    } else {
-        strncpy(fn, symbol, bufSize - 1);
-        fn[bufSize - 1] = '\0';
-    }
-}
-
-#define D(X) { \
-  void* callstack[2]; \
-  int frames = backtrace(callstack, 2); \
-  char** symbols = backtrace_symbols(callstack, frames); \
-  char fn[512]; \
-  extractc_fn(symbols[1], fn, 512); \
-  std::cout << getpid() << "/ " << gettid() << ": [" << fn << "->" <<__func__ << "] " << __LINE__ << ": " << X << std::endl; \
-  free(symbols); \
-}
+#include "caret_trace/DEBUG.hpp"
 
 #define SELECT_NODES_ENV_NAME "CARET_SELECT_NODES"
 #define IGNORE_NODES_ENV_NAME "CARET_IGNORE_NODES"
@@ -154,11 +125,7 @@ TracingController::TracingController(bool use_log)
   ignored_topic_names_(get_env_vars(IGNORE_TOPICS_ENV_NAME)),
   ignored_process_names_(get_env_vars(IGNORE_PROCESSES_ENV_NAME)),
   select_enabled_(selected_topic_names_.size() > 0 || selected_node_names_.size() > 0),
-  ignore_enabled_(
-    ignored_topic_names_.size() > 0 ||
-    ignored_node_names_.size() > 0 ||
-    ignored_process_names_.size() > 0
-  ),
+  ignore_enabled_(ignored_topic_names_.size() > 0 || ignored_node_names_.size() > 0),
   use_log_(use_log)
 {
   if (select_enabled_ || ignore_enabled_) {
@@ -181,10 +148,15 @@ TracingController::TracingController(bool use_log)
     if (ignored_topic_names_.size() > 0) {
       info(IGNORE_TOPICS_ENV_NAME + std::string(": ") + get_env_var(IGNORE_TOPICS_ENV_NAME));
     }
-    if (ignored_process_names_.size() > 0) {
-      info(IGNORE_PROCESSES_ENV_NAME + std::string(": ") + get_env_var(IGNORE_PROCESSES_ENV_NAME));
-    }
   }
+
+  if (ignored_process_names_.size() > 0) {
+    info(IGNORE_PROCESSES_ENV_NAME + std::string(": ") + get_env_var(IGNORE_PROCESSES_ENV_NAME));
+  }
+
+  is_ignored_process_ =
+      ignored_process_names_.size() > 0 &&
+      partial_match(ignored_process_names_, std::string(program_invocation_short_name));
 
   check_condition_set(selected_node_names_, use_log);
   check_condition_set(ignored_node_names_, use_log);
@@ -627,17 +599,7 @@ bool TracingController::is_allowed_client_handle(const void * client_handle) {
 
 bool TracingController::is_allowed_process()
 {
-  D(0)
-  if (ignore_enabled_) {
-    auto is_ignored_process = partial_match(
-      ignored_process_names_,
-      std::string(program_invocation_short_name)
-    );
-    if (is_ignored_process && ignored_process_names_.size() > 0) {
-      return false;
-    }
-  }
-  return true;
+  return !is_ignored_process_;
 }
 
 std::string TracingController::to_node_name(const void * callback)
